refactor: move api general details in configuration entry
