+++ gravitee-apim-console-webui/src/management/api/general-info/api-general-info-duplicate-dialog/api-general-info-duplicate-dialog.component.scss
