+++ gravitee-apim-console-webui/src/management/api/general-info/api-general-info-quality/api-general-info-quality.component.scss
