+++ gravitee-apim-console-webui/src/management/api/general-info/api-general-info.component.html
